feat: HEAD 메소드 처리 함수 일부 구현
변경점
- HEAD 메소드 처리 함수 일부 구현
- setConfig 함수 오류 수정 및 makeMsg 직전에 setConfig 추가
- Config 파싱 시 index와 error_page를 전체 경로로 저장하도록 수정
- HTTP 양식에 맞춘 timestamp 처리 함수 추가

+++ includes/Http.hpp
@@ -20,11 +20,12 @@ struct StartLineReq : public StartLine
 
 struct StartLineRes : public StartLine
 {
-	uint8_t		status_code;
+	uint16_t		status_code;
 };
 
 class Http
 {
+	// TODO: map으로 바꾸기
 	std::multimap<std::string, std::string>	_headers;
 	std::deque<uint8_t>						_body;
